refactor: rename UiLink to UiLabel
